test detection of grouping changes
